pyproject.toml
[project]
name = "tinyloop"
version = "0.1.0"
description = "A super lightweight library for LLM-based applications"
readme = "README.md"
authors = [{ name = "Fernando Meira", email = "fmeira.filho@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "instructor>=1.10.0",
    "litellm>=1.75.9",
    "mlflow>=3.3.1",
    "pydantic>=2.11.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
]
